Update dependencies to use optimistic versioning
This will enable us to test future Rails and DelayedJob versions
without bumping TrackedJob contstraints.

This is a tradeoff as TrackedJob could now break if its dependencies
are updated in an incompatible way.

We therefore rely on the TrackedJob consumer to coordinate the
versions.

If a breaking change is found we should:
* add pessimistic versioning
* fix the breaking change
* add back optimistic versioning